+++ changelog.md
@@ -1,6 +1,7 @@
 ## unreleased
 - Update dependencies, as a result of `Sirupsen/logrus` -> `sirupsen/logrus` (#333)
 - Add a Docker subcommand (#335)
+- Ensure repository names are always lowercase (#338)
 
 ## v1.0.965 (2017-08-23)